refactor(forms): split WasteSubmissionForm into modular subcomponents and extract validation helpers (#1045)
